routes abstracted to its own modules for all model endpoints
